+++ test/switchboard/native/PolygonL1Switchboard.t.sol
@@ -74,7 +74,7 @@ contract PolygonL1SwitchboardTest is Setup {
 
     function deployContractsOnSingleChain(
         ChainContext storage cc_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256[] memory transmitterPrivateKeys_
     ) internal {
         // deploy socket setup
@@ -164,7 +164,7 @@ contract PolygonL1SwitchboardTest is Setup {
 
     function addPolygonL1Switchboard(
         ChainContext storage cc_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256 capacitorType_
     ) internal returns (SocketConfigContext memory scc_) {
         polygonL1Switchboard = new PolygonL1Switchboard(
@@ -191,7 +191,7 @@ contract PolygonL1SwitchboardTest is Setup {
         ChainContext storage cc_,
         address deployer_,
         address switchBoardAddress_,
-        uint256 remoteChainSlug_,
+        uint32 remoteChainSlug_,
         uint256 capacitorType_
     ) internal returns (SocketConfigContext memory scc_) {
         vm.startPrank(deployer_);
